requirements.txt
beautifulsoup4==4.12.3
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
mysql-connector-python==9.2.0
oauthlib==3.2.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
soupsieve==2.6
uritemplate==4.1.1
urllib3==2.3.0
